Extract date and SQLite helper functions
