+++ util/task.py
@@ -2,19 +2,16 @@ import json
 import os
 from os import path
 import time as stdlib_time
-from datetime import datetime, timedelta, date, time
+from datetime import datetime, timedelta
 import logging.config
 from discovergy.discovergy import Discovergy
 import redis
-from models.baseline import BaseLine
-from models.pkv import PKV
-from models.savings import UserSaving, CommunitySaving
 from util.error import exception_message
-from util.energy_saving_calculation import estimate_energy_saving_each_user,\
-    estimate_energy_saving_all_users, get_all_user_meter_ids, calc_energy_consumption_last_term
 from util.database import create_session
-from util.pkv_calculation import define_base_values, calc_pkv
-from util.sqlite_helpers import get_all_meter_ids, get_all_users
+from util.date_helpers import calc_support_year_start, calc_term_boundaries,\
+    calc_end, calc_support_week_start, calc_two_days_back
+from util.sqlite_helpers import get_all_meter_ids, write_baselines,\
+    write_savings, write_base_values_or_pkv
 
 
 log_file_path = path.join(path.dirname(
@@ -28,253 +25,6 @@ redis_host = os.environ['REDIS_HOST']
 redis_port = os.environ['REDIS_PORT']
 redis_db = os.environ['REDIS_DB']
 last_data_flush = None
-message_timestamp = datetime.strftime(datetime.utcnow(), '%Y-%m-%d %H:%M:%S')
-
-
-def calc_term_boundaries():
-    """ Calculate begin and end of previous and ongoing terms.
-    :return: begin and end of the previous and the current support year until
-    today as unix milliseconds timestamps
-    :rtype: tuple(int)
-    """
-
-    # Calculate timestamps for ongoing term
-    begin_ongoing_term = calc_support_year_start()
-    end_ongoing_term = round(datetime.combine(datetime.now().date(), time(0, 0,
-                                                                          0)).timestamp()
-                             * 1000)
-
-    # Calculate timestamps for previous term
-    end_prev_term = round((datetime.fromtimestamp(
-        begin_ongoing_term/1000) - timedelta(days=1)).timestamp() * 1000)
-
-    previous_year = datetime.fromtimestamp(
-        begin_ongoing_term/1000).year - 1
-    begin_previous_term_date = datetime.fromtimestamp(
-        begin_ongoing_term/1000).date().replace(year=previous_year)
-    begin_prev_term = round(datetime.combine(begin_previous_term_date,
-                                             time(0, 0, 0)).timestamp() *
-                            1000)
-
-    return begin_ongoing_term, end_ongoing_term, begin_prev_term, end_prev_term
-
-
-def calc_end():
-    """ Calculate timestamp of end of interval. """
-
-    # Multiply the result of timestamp() from the standard library by 1000 and
-    # round it to have no decimal places to match the timestamp format required
-    # by the discovergy API
-    return round(datetime.utcnow().timestamp() * 1000)
-
-
-def calc_support_year_start():
-    """ Calculate start of BAFA support year.
-    :return:
-    March, 12th the year before if today is between January, 1st and March, 11th
-    March, 12th of the current year otherwise
-    :rtype: unix timestamp in milliseconds
-    """
-
-    now = datetime.utcnow()
-    start_day = 12
-    start_month = 3
-    if (now.month < start_month) or (now.month == start_month and now.day <
-                                     start_day):
-        start_year = now.year - 1
-    else:
-        start_year = now.year
-    d = date(start_year, start_month, start_day)
-    t = time(0, 0)
-    support_year_start = round(datetime.combine(d, t).timestamp() * 1000)
-    return support_year_start
-
-
-def calc_support_year_start_datetime():
-    """ Calculate start of BAFA support year.
-    :return:
-    March, 12th the year before if today is between January, 1st and March, 11th
-    March, 12th of the current year otherwise
-    :rtype: datetime.date in UTC
-    """
-
-    now = datetime.utcnow()
-    start_day = 12
-    start_month = 3
-    if (now.month < start_month) or (now.month == start_month and now.day <
-                                     start_day):
-        start_year = now.year - 1
-    else:
-        start_year = now.year
-    d = date(start_year, start_month, start_day)
-    t = time(0, 0)
-    return datetime.combine(d, t)
-
-
-def calc_support_week_start():
-    """ Calculate start of BAFA support week.
-    :return:
-    March, 12th of the current year if today is between March, 12th
-    and March, 18th
-    7 days back in time otherwise
-    :rtype: unix timestamp in milliseconds
-    """
-
-    now = datetime.utcnow()
-    if (now.month == 3) and (11 < now.day < 19):
-        d = date(now.year, 3, 12)
-    else:
-        d = (now - timedelta(days=7)).date()
-    t = now.time()
-    support_week_start = round(datetime.combine(d, t).timestamp() * 1000)
-    return support_week_start
-
-
-def calc_two_days_back():
-    """ Calculate timestamp of 48 hours back in time. """
-
-    # Multiply the result of timestamp() from the standard library by 1000 and
-    # round it to have no decimal places to match the timestamp format required
-    # by the discovergy API
-    return round((datetime.utcnow() - timedelta(hours=48)).timestamp() * 1000)
-
-
-def write_savings(session):
-    """ Write the energy savings of each user and the community to the
-    SQLite database.
-    """
-
-    start = calc_support_year_start_datetime()
-    try:
-        for key, value in estimate_energy_saving_each_user(start,
-                                                           session).items():
-
-            if value is None:
-                message = """Cannot calculate saving for meter id {} on {}
-                because last term\'s energy consumption or estimated energy
-                consumption is missing. Writing 0.0 instead""".format(key,
-                                                                      message_timestamp)
-                logger.info(message)
-
-                # Create UserSaving instance and set saving to 0.0
-                user_saving = UserSaving(datetime.utcnow(), key, 0.0)
-
-            else:
-                user_saving = UserSaving(datetime.utcnow(), key, value)
-
-            session.add(user_saving)
-
-        # Create CommunitySaving instance
-        community_saving = estimate_energy_saving_all_users(start, session)
-        session.add(CommunitySaving(datetime.utcnow(), community_saving))
-
-        session.commit()
-    except Exception as e:
-        message = exception_message(e)
-        logger.error(message)
-
-
-def write_baselines(session):
-    """ Write the baseline for each user to the SQLite database. """
-
-    start = calc_support_year_start_datetime()
-    try:
-        for meter_id in get_all_user_meter_ids(session):
-            baseline = calc_energy_consumption_last_term(meter_id, start)
-
-            if baseline is None:
-                message = """Cannot write baseline for meter id {} on {} because last term\'s energy
-                consumption is missing""".format(meter_id, message_timestamp)
-                logger.info(message)
-
-            else:
-                # Create BaseLine instance
-                session.add(BaseLine(datetime.utcnow(), meter_id, baseline))
-
-        session.commit()
-    except Exception as e:
-        message = exception_message(e)
-        logger.error(message)
-
-
-def write_base_values_or_pkv(session):
-    """ If yesterday was the start of the support year, write the base values
-    for all users to the SQLite database.
-    Otherwise, write yesterday's pkv for all users to the SQLite database.
-    :param sqlalchemy.orm.scoping.scoped_session session: the database session
-    """
-
-    yesterday_date = (datetime.today() - timedelta(days=1)).date()
-    today_date = datetime.today().date()
-    _time = time(0, 0, 0)
-    yesterday = datetime.combine(yesterday_date, _time)
-    today = datetime.combine(today_date, _time)
-    support_year_start = calc_support_year_start_datetime()
-
-    if today == support_year_start:
-        write_base_values(yesterday, session)
-    else:
-        write_pkv(yesterday, session)
-
-
-def write_base_values(dt, session):
-    """ Write the base values for each user to the SQLite database.
-    :param datetime dt: the date to write the values for
-    :param sqlalchemy.orm.scoping.scoped_session session: the database session
-    """
-
-    for user in get_all_users(session):
-        base_values = define_base_values(user.inhabitants, dt)
-
-        # Create PKV instance
-        session.add(PKV(dt, user.meter_id, base_values['consumption'],
-                        base_values['consumption_cumulated'],
-                        base_values['inhabitants'], base_values['pkv'],
-                        base_values['pkv_cumulated'], base_values['days'],
-                        base_values['moving_average'],
-                        base_values['moving_average_annualized']))
-
-    session.commit()
-
-
-def write_pkv(dt, session):
-    """ Write the pkv for each user to the SQLite database.
-    If for one user there are no yesterday's values in the database, write the
-    base values for that user.
-    If today's entry already exists for a user, skip writing that entry.
-    :param datetime dt: the date to write the values for
-    :param sqlalchemy.orm.scoping.scoped_session session: the database session
-    """
-
-    for user in get_all_users(session):
-
-        try:
-            # Check if entry exists
-            pkv_today = session.query(PKV).filter_by(
-                date=dt, meter_id=user.meter_id).first()
-
-            # Create today's entry if it does not exist
-            if not pkv_today:
-                dataset = calc_pkv(
-                    user.meter_id, user.inhabitants, dt, session)
-
-                # If there are no yesterday's values in the database for this user,
-                # define the base values
-                if dataset is None:
-                    dataset = define_base_values(user.inhabitants, dt)
-
-            # Create PKV instance
-                session.add(PKV(dt, user.meter_id, dataset['consumption'],
-                                dataset['consumption_cumulated'],
-                                dataset['inhabitants'], dataset['pkv'],
-                                dataset['pkv_cumulated'], dataset['days'],
-                                dataset['moving_average'],
-                                dataset['moving_average_annualized']))
-        except Exception as e:
-            message = exception_message(e)
-            logger.error(message)
-
-    session.commit()
 
 
 def check_and_nullify_power_value(reading, meter_id):
